fix(parser): create conditional sequence
