+++ Makefile
@@ -27,8 +27,8 @@ DEBUG   := scanning.c parsing.c
 HEADERS := minishell.h hash_table.h builtins.h parser.h parser.h tree.h
 SOURCES := minishell.c hash_table.c hash_table_utils.c
 SOURCES += sig_setup.c sig_events.c echo.c
-SOURCES += scanner.c scanner_utils.c token_word.c parser.c error.c
-SOURCES += rules1.c rules2.c rules3.c rules4.c tree.c tree_utils.c table.c
+SOURCES += scanner.c scanner_utils.c token_word.c parser.c error.c table.c
+SOURCES += rules1.c rules2.c rules3.c rules4.c rules5.c tree.c tree_utils.c
 
 SOURCES += $(DEBUG)
